Move code to shipInit.
